requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
async-timeout==4.0.2
attrs==23.1.0
binance-connector==3.1.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.1.0
cycler==0.11.0
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
fonttools==4.40.0
frozenlist==1.3.3
idna==3.4
# Editable Git install with no remote (iribot==1.0.0)
-e /Users/iri369/code/python/iri
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
Mako==1.2.4
MarkupSafe==2.1.3
matplotlib==3.7.1
multidict==6.0.4
numpy==1.25.0
openai==0.27.8
packaging==23.1
pandas==2.0.2
pandas-ta==0.3.14b0
Pillow==9.5.0
psycopg2-binary==2.9.6
pycryptodome==3.18.0
pyparsing==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.17
tqdm==4.65.0
typing_extensions==4.7.0
tzdata==2023.3
urllib3==2.0.3
websocket-client==1.6.1
Werkzeug==2.3.6
yarl==1.9.2